DonatePrice type changed from decimal to double, changed AddChallengeAsync method

+++ DonateChallenger/Services/ChallengeCatalog/ChallengeCatalog.API/Repositories/Abstractions/IChallengeRepository.cs
@@ -5,7 +5,7 @@ namespace ChallengeCatalog.API.Repositories.Abstractions;
 
 public interface IChallengeRepository
 {
-    Task<long?> AddChallengeForStreamerAsync(string description, decimal donatePrice, string donateFrom, string streamerId, string? title);
+    Task<long?> AddChallengeForStreamerAsync(string description, double donatePrice, string donateFrom, string streamerId, string? title);
     Task<PaginatedChallenges> GetPaginatedCurrentChallengesAsync(int currentPage, int challengesPerPage, string streamerId, int? minPriceFilter = null, int? sortByCreatedTime = 0, bool sortBySkipped = false, bool sortByCompleted = false);
     Task<bool?> UpdateChallengeStatusByIdAsync(long challengeId, bool skipped, bool completed);
     Task<ChallengeEntity?> GetChallengeByIdAsync(long challengeId, bool loadRelatedEntities = false);
